backend/requirements.txt
# Django Core
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0

# Filtering
django-filter==23.5

# Documentation
drf-spectacular==0.27.0

# Tree structure for accounts
django-mptt==0.14.0

# Reports
WeasyPrint==60.2
openpyxl==3.1.2
Jinja2==3.1.3

# Audit
django-simple-history==3.5.0
